requirements.txt
asgiref==3.5.2
django-phonenumber-field==7.0.1
Django==4.1.4
djangorestframework==3.14.0
phonenumberslite==8.13.1
pip==21.3.1
pytz==2022.6
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.1
